chore: clippy to zero warnings (pre-existing lint debt) (#95)
cargo clippy --all-targets was 19 lib warnings + 1 denied never_loop
error. All pre-existing (main was never lint-clean after the classic-cards
roadmap acceptance item); no behavior change — verified by the full suite.

- never_loop: intercept_counter_secrets (secret.rs) — the loop body always
  returns (only the first counter-secret fires, HS rule); rewritten as
  if-let over secrets.into_iter().next(), identical semantics
- unused vars: _-prefixed params (resolve_swap_attack_health,
  resolve_full_heal_and_taunt, resolve_gain_stats_and_taunt), MortalStrike
  pattern, test locals
- too_many_arguments x3: #[allow] + param-list comment (repo convention)
- needless_borrow: card.name / CardId.0 are &str — pass by value
- collapsible_if, needless mut, useless cast, empty print literal,
  bool assert_eq! -> assert!, unused imports, missing docs (EntityView /
  redirect_attack / aura fields)
- non_snake_case test helpers PlayerId1/2: allow (392 call sites — the
  rename would dwarf the cleanup)

cargo test: 419 green. cargo clippy --all-targets: 0 warnings, 0 errors.

+++ src/rl/views.rs
@@ -63,18 +63,27 @@ pub struct EntityView {
     pub has_trigger: bool,
     /// Battlecry/spell magnitudes (M5 card text — v6 A_TEXT idea)
     pub bc_damage: i32,
+    /// Cards drawn
     pub bc_draw: i32,
+    /// Minions summoned
     pub bc_summon: i32,
+    /// Attack/health granted
     pub bc_buff: i32,
+    /// Health restored
     pub bc_heal: i32,
+    /// Characters frozen
     pub bc_freeze: i32,
+    /// Minions destroyed
     pub bc_destroy: i32,
     /// Deathrattle magnitudes
     pub dr_damage: i32,
+    /// Cards drawn
     pub dr_draw: i32,
+    /// Minions summoned
     pub dr_summon: i32,
-    /// Aura magnitudes
+    /// Aura attack bonus
     pub aura_attack: i32,
+    /// Aura health bonus
     pub aura_health: i32,
 }
 
@@ -468,7 +477,7 @@ mod tests {
         assert_eq!(views.len(), plain.len());
         assert_eq!(views[0].kind, "end_turn");
         for v in &views {
-            assert_eq!(v.index as usize, views[v.index].index);
+            assert_eq!(v.index, views[v.index].index);
             assert!(!v.description.is_empty());
         }
     }
